(Hacky) :bow: allow to change mysql queries before execution
`ANY` and arrays behave differently in MySQL and Postgres.

  In Postgres:
  ```sql
  where x = ANY ('{1,2,3,4}'::integer[])
  ```

  In MySQL:
  ```sql
  where x = ANY (1,2,3,4)
  ```
We do need to write the query in a postgres style though, because the mysql
version doesn't type check.  This introduces a function to modify the query
before it's executed with MySQL.Simple.  We simply replace the things we don't
like from the postgres version.  This is pretty sad and ugly, but I couldn't
figure out a different solution. We should remove this function if we figure
out a better way to do this.