Inline `Query.execute` function
The way to run a query is different between different databases, and so
we pay a price trying to catch this in a single abstraction. We have to
pass in the function that actually does the work. As we add tracing to
the mix, which requires additional know-how about the internals of the
query type for MySQL and Postgres, there's yet more configuration
functions we have to pass in to allow the generic `execute` function to
do its job for both databases.

Ultimately `execute` has only two consumers: MySQL and Postgres. The
amount of configuration required for it to handle these use cases does
not seem worth the gain. Inlining this function results in hardly any
dupplication being introduces, but does remove a bunch of code.