convert `= ANY` to `IN`
MySQL doesn't support `ANY`, we can use `IN` instead.
We need to write the query with `ANY` though, because postgres-typed is
expecting that. The code won't compile otherwise.

We use this library because it has some nice benefits to writing string-based
regex. The template haskell creates a `Traversal` and we can use `lens` to
view and change the matches. Additionally, we get compiler errors if the regex
doesn't compile.

Reference: https://chrispenner.ca/posts/lens-regex-pcre