Parser for query highlights
To be able to see slow queries in the new relic dashboard we need to
send distinguishing information for queries. We're already passing New
Relic the entire query, but it seems to only use that information for
slow queries.

This commit adds a parser that attempts to retrieve the table queried
and the SQL operation used. It seems to work well enough with a
selection of the queries in this repo. We might need to tweak it further
in the future.