Simplify `Query` type
For some time now we've worked with this query type:

  PGQuery q a => Query q

But who really understands what that `q` means? Not me! Answering that
question requires a deep dive into the `postgresql-typed` package.
Usually we don't really need to know what it is and that is sort of
fine, but we can do better.

This changes the query type to this:

  Query row

Note that the `PGQuery` constraint is gone, we don't need it anymore.
`row` signifies the type of row returned by the query. So if you do a
`SELECT hat count(*)`, row will be something like `(Hat, Int)`. I feel
that makes a great deal more sense!