All log functions take additional optional context
In preparing the next application to support the new `Logger` module, I
noticed that one thing Rollbar allows is for additional context values
to be passed to a logging function. This adds that same support to all
our logging functions. Doesn't seem a huge burden for the places where
it isn't wanted, because all you need to type is an empty list.

Plus, we use the opportunity to write some examples for some functions
in `Log`.