Mark severity of all our health checks
Some health checks should bring the service down when they fail, in the
hope that this restart might resolve the connectivity issue. But this
treatment should only be given to external services that are critical to
the apps functioning, such as the database.

When monitoring and logging external services go down, like Rollbar or
New Relic, we should keep the app live. Although we should look into
fixing our monitoring problems there's no reason to create trouble for
our users.