Add the query template to the log context
At the moment we're explicitly logging the query template, but we're not
adding it to the logging context. The logging context only contains the
full query string, but we wrap it in a secret because it might contain
sensitive data and so it's not very useful in our logs.

By adding the template uncaught exceptions while querying will contain
information about which query failed.