Inline creation of database connection
I'd like to try replace our current `mysql-simple` library with
`mysql-haskell-persistent`. `mysql-haskell-persistent` has its own
support for creating a database connection pool, which means the process
of creation a `Connection` type for it is different then creating one
for Postgres. This means the `GenericDb` abstraction for creating a
connection cannot be used here.

We remove it by inlining that code into the MySQL and Postgres modules
directly. This helps us get rid of a `ConnectionPoolConfig` type, which
is a nice bonus.