Migrate functions from `Log` to `Platform`
`Log` was exposing the regular functions for logging in our
applications, and also some types intended for working with the log
handler in our libraries. For example, the `database` package adds some
special logging metadata containing query details, and our `webserver`
package contains some logic for initializing the log handler.

This commit keeps the basic logging API in the `log` module, but moves
all the other helpers into the `Platform` module. This is an attempt to
make the `Log` module super approachable (we want to encourage people to
use it!), and putting all helpers for library building in `Platform`.