Add `Expect` type to `http` library
I noticed health check requests to Rollbar were failing in the decoding
step. Looking at the endpoint we're calling on Rollbar I noticed it
returned a plaintext response. We couldn't decode this as JSON.

We were using the `Anything` type to indicate that we were ok with any
response and it has a JSON decoder that always succeeds, but even that
decoder expects the response to be valid JSON to begin with. That wasn't
the case here, since we were getting a plain text (without any
surrounding quotes which would turn it into a JSON string).

This commit takes an approach from the `elm-http` handbook to deal with
this. It defines an `Expect` type that can be used to indicate the
desired format of the response.