Explicit imports of Haskell `base` dependencies
We used to get a lot of Haskell functions and types from `Nri.Prelude`,
which in turn was piping them through from `Protolude`. The selection of
Haskell types we piped through has no particular rhyme or reason to it.
When we created `Nri.Prelude` we simply took stock of which Haskell
types and functions we used and piped those through. That way, we
thought, at least we would put a lid on our use of Haskell types and
functions: we could use what we were already using but wouldn't extend
our usage further.

Now it's time to take a further step. We'd like to open source
`Nri.Prelude` at some point (although under a different name), and
because it is intended to be an Elm-inspired replacement Prelude for
Haskell having a ton of `Protolude` functions in there doesn't make much
sense.

We can have a debate about exactly which non-Elm-like Haskell functions
and types add enough value that they warrant exceptional inclusion in
`Nri.Prelude`, but that's one bike shed we'd rather not paint. So we
took a more drastic step in not exposing any non-Elm-like Haskell
functions or types, except a couple of classes our `Basics` module uses
like `Eq and Ord`.

The implications of this change is that in the rest of our codebase, if
we want to use a non-Elm-like Haskell function or type we'll have to get
it explicitly from whatever `base` module it lives in. In other words:
we have to add a boatload for import statements. Arguably, this isn't
great.

In this commit we do so anyway because, as mentioned previously, it will
allow us to move forward with opensourcing `nri-prelude`. There's
another benefit in that these explicit imports are providing a lot of
visibility on where our code isn't as Elm-like as we might want. This
shows opportunties for small refactors that remove Haskell-specific
functions. This commit doesn't make all those refactors because it would
be too much work to do all at once. To help us make these refactors when
we have an opportunity to do so, we now have the following guide:

https://paper.dropbox.com/doc/Cherry-flavored-refactors--AqjxGugkzx5p3ubrzbFUZzzxAg-cTNnf1OD1i6sLAPY2LBbC