Fail requests on non-200 status codes
Weirdly the parse request function you use determines the behavior of
the library when a request returns with a non-200 status code. From the
library docs:

    A Request created by this function won't cause exceptions on non-2XX
    response status codes. To create a request which throws on non-2XX
    status codes, see parseUrlThrow

Although we're handling an exception for error status codes, that
exception would never be thrown. Instead a failing request would like
end up in the happy flow and then sometimes fail decoding. This commits
fixes the issue.