Trace http requests from third party libraries
In previous commits we added support for tracing http requests to all
requests made directly using our `http` package. We also ensures these
requests now by default get an `x-request-id` set. This we could do
without making any API changes.

Unfortunately this doesn't cover all outgoing http requests, or even
most of them. A pattern that we see a lot is that of a third party
library making HTTP requests on our behalf. Such a library often takes a
`Manager`, and the `http` library provides a function for obtaining such
a `Manager`.

We'd like this `Manager` we pass out to have the same tracing and
`x-request-id` capability that we have available for requests made
directly through the `http` library. Making these modifications to this
manager is possible, but requires access to the log handler, which we
only have inside of a `Task`. So for this we need to make a API-breaking
change: our `withThirdParty` function can now only run inside of a
`Task` to ensure a log handler is available.