Simplify `database` API
The `Postgres` and `MySQL` modules from the `database` package were
exporting a number of aliases for the same thing: `doQuery`, and
`getMany` would both launch a query returning multiple rows. `getOne`
and `modifyExactlyOne` would do the same thing, then throw an error if
the result wasn't exactly a single row.

In this simpler api we only have `doQuery` left as a function to query
with. It is a drop in replacement for `getMany`. For the places where we
are using `getOne` and `modifyExactlyOne`, we can use `doQuery` and then
subsequently calling a filter that either selects the one row result if
that's what we get, or throws an exception otherwise.

The purpose of this change is two-fold. First it makes the API smaller
and removes some duplication and potential source of confusion as to
which function to use for querying. Secondly, by making the 'expect a
single row' part of the API a separate function, the one remaining query
function `doQuery` now returns a `Task e [row]`. The error parameter `e`
is free right now, which allows us to put an SqlException in there in a
future commit.

This commit doesn't make the necessary changes in the projects that
depend on `database`. We follow up with these in a separate commit.