Move `expectOne` to `webserver` package
The `expectOne` helper ensures a list of queried rows contains a single
element. If so it returns that element, if not an error. That error type
we then have to handle in the calling code to turn it into an
`ErrorResponse` that we can log / return to the user.

In theory we could have some custom error handling logic in place that
acts on the knowledge that a query returns an unexpected 0 or more than
1 row with some special logic. In practice we're not writing any such
logic. So for all currently known use cases, it would be way more
convenient if `expectOne` immediately returns an `ErrorResponse`,
because then we can save ourselves the effort of constructing one
manually.

That's what this commit does. Because the `ErrorResponse` type lives in
the `webserver` package, we need to move our `expectOne` function there,
or take a dependency on `webserver` in `database`. The latter seems
backwards, so we prefer to move the helper function.

If in the future we do want to write that custom error handling logic we
still can, we just have to implement a manual function to pattern match
on the rows returned, and then do something depending on the amount of
rows we get. Since we're not doing that so far, the loss of a helper
function for this use case doesn't seem a big deal.

We do loose some contextualization. Our generic `expectOne` function
doesn't know what we were querying, and so won't be able to provide any
context about that in the errors it creates. This is problematic, but
will be addressed in a follow-up commit: The query context is available
to the code running the query, and if we perform our error handling in
that context as well we'll be able to add it to any errors created.