`doQuery` takes functions for handling response
This changes the API of `Postgres.doQuery` and `MySQL.query`. Before we
might write this:

    Postgres.doQuery
      connection
      [Postgres.sql|
        ...
      |]
    |> andThen WebServer.expectOneRow

Now we write this:

    Postgres.doQuery
      connection
      [Postgres.sql|
        ...
      |]
      WebServer.expectOneRow

That is, the function for handling the response rows has become an
explicit argument of `doQuery`. As a result we always need to pass one,
even if we don't expect a single row. To this end we add a
`WebServer.expectQuerySuccess` function to use in cases where previously
we wouldn't have an `expectOneRow`.

I'm not sure about argument order. I think this is more natural, with
the query coming first and then the function that does something with
the resulting rows, but don't feel super strongly.

So why this change? Two reasons:

1. By passing the response handler function to the query itself we can
   invoke it in the log context of the query. That means that if it
   fails, for example when the query returns 0 or more than 1 row, the
   exception we produce gets all the query info available on the
   context, like the query that was ran.
2. At the moment `WebServer.expectQuerySuccess` isn't really doing
   anything, and so it seems superfluous. But a goal is to have
   `doQuery` return explicit errors when queries fail because of, for
   example, timeouts. We can now add this logic and immediately add
   handling logic for those errors in `expectQuerySuccess`.

Which isn't to say we are forced to pass either `expectOneRow` or
`expectQuerySuccess` to all `doQuery` calls. We can put a custom
function in there if we feel we need to add some custom error handling
or logging. For most cases though `expectOneRow` and
`expectQuerySuccess` should be fine.