Make `Environment.decode` throw
Generally speaking I like the fact that the `Environment.decode`
function returns an explicit exception rather than throwing it. I feel
that's the design we should aspire too for all our APIs. Unfortunately
what happens next is that consumers for `Environment.decode` throw the
error instead. They don't have any other options because they run in an
`IO` or `Acquire` context.

This commit is part of an effort to remove our use of exception
throwing, then putting in place some hlint rules to prevent backsliding.
By making an exception to this new rule not to throw in the case of the
`Environment` module, we win not having to make that exception in a lot
of other places in our application code.

On the long run I think we can remove even the exception for
`Environment` by having the settings parsing logic of our apps run in
`Task` rather than `IO`. This would allow our `Environment.decode`
function to return a `Task` as well which has a parameter for an
explicit exception type.