Support non-JSON bodies
We're currently assuming all request bodies to contain JSON. This means
that even for GET requests that don't need to contain a body, we're
sending along a JSON encoded `()` value. Aeson encodes this as an empty
list: `[]`.

We've started noticing this because requests made to check Rollbar
uptime were failing with a 400 status code. Rollbar fails if you send
messages to its /ping endpoint containing a json body.

This commit addresses the issue by taking the same approach to
supporting different request bodies that Elm's `http` library does.