Let services decide which checks are fatal
Before each library would define whether a failing health check meant
that the service relying on it was fatal to that service. This doesn't
work out.

For example: usually if the monolith isn't available for authentication
that's a fatal problem. But it shouldn't be for the tutorials service,
because even without authentication it is still able to serve tutorials
to students.