Backport escaping of inserts in MySQL
Working on the quiz-engine we discovered MySQL inserts would sometimes
fail because interpolated data was escaped improperly. This is down to
a difference in how postgresql-typed escapes text columns for Postgresql
and the way MySQL expects text columns to be escaped. This commit
ensures text-like columns are escaped using the correct MySQL strategy.

MySQL has an option ANSI QUOTES that will make it compatible with ANSI
escaping rules, basically adopting the same strategy Postgres has for
escaping. Enabling this option might be the nicer way to fix this
problem, but it's unclear at this moment if that's a safe thing to do
with our main MySQL database. If we do flip that switch we can remove
this hack.