No need for MySQL.Single type
Previously when a MySQL query in Haskell returned a row with a single
column you'd get back a type wrapped in a `Single` constructor. Now
that wrapper is gone! This makes the API for the `MySQL` module for
queries the same as the one of the `Postgres` module, which didn't have
this wrapper type either.