Allow libraries to test against the database
Up to this point only project could write tests that talked to the
database. This commit makes it so libraries can to, using a shared
`libry_tests` database.

We also set up the `database` package to make use of this new
functionality, in preparation of some tests that will be added in
upcoming commits.