Make MySqlQueryable do less
The purpose of `MySqlQueryable` is to make it possible to use `doQuery`
with two different types of queries. Reads that return a list of row
types and writes that return `()`.

The implementation of the class covered that plus a whole ton of
additional query processing logic that was common for both use cases. To
not have a lot of duplication the two instances of the class both called
the large function doing all the stuff and passed in a different helper
for doing the actual query.

This commit flips that around. The class just does the tiny query task
for which each instance defines separate behavior. That saves us having
to pass around a bunch of complicated functions.