Be more upfront of errors we ignore
We have a type signature for `transaction` like this:

```
transaction :: Connection -> Task e a -> Task e a
```

This is a bit too good to be true. The transaction is a database
operation that can fail for all sorts of reasons (for example: the
database being down). But we don't see that in the signature above.

We achieve this by throwing any database errors as runtime exceptions.
This change makes that a little bit more visible in the code, will give
slightly nicer errors if transaction exceptions do occur, and paves the
road for cleaning up the code a bit further in subsequent commits.