Test MySQL queries with question marks
This is a regression test for a problem seen in production. When we
perform an MySQL insert from a Haskell service and one of the inserted
columns contains a question mark then the query fails with a
`WrongParamsCount` error.

It looks like the internal library we're using for making queries to
MySQL things the question mark is a placeholder it should fill in, only
that isn't the case here.