Move tests into test/ directory
Currently the tests are part of the package source code. That means
dependencies of this library, such as the `webserver` library, also need
to be able to compile the tests. The tests use the query quasi-quoter
and need a database, and so every dependency of this package also
requires a database to compile against.

By moving the test source files into a separate directory they stop
being part of the package proper. That way dependencies don't need to be
able to compile them.