Simplify query logging
We were doing a lot of string manipulation for logging queries. We've
seen in benchmark tests that these string manipulations for long queries
can have a big impact on performance.

Instead we log a simpler message without doing any query manipulation
from within the query context. The query string will be set on the
context and so logged as metadata as well (but unformatted).