Don't define custom MySQLColumn instances
The intent for the PGColumn instance that `postgresql-typed` gives us
and the `MySQLColumn` instance that we model after it is that it
describes safe conversions of database types into Haskell types. It's
intentionally not a decoder with a failure mode.

If we try to shoehorn decoding operations into it we have to resort to
mechanisms like `impureThrow`. That's not a great pattern to embrace in
our code base.

So to prevent ourselves to go down this road we don't expose
`mysqlDecode`, preventing us from defining custom `PGColumn` instances.
We can derive them on newtypes, which is fine and safe. If we want to
read data from the database and transform it into other data in ways
that can fail we can still do so, but not as part of MySQL parsing
logic.