Remove readiness checks for reporters
We've had these readiness for a while, but their use has been dubious at
best. They're part of the reason the readiness test is slower, and it's
not all that clear what they're telling us we don't already know: if
reporting isn't working we probably know because we're not getting any
information.

Removing these for now. Some error logging (to file) in case reporting
to Bugsnag or NewRelic fails might be more useful.