Limit calls for monotonic time
Katip internally does the same thing. Getting the current time and the
start and end of every span can start having a significant impact on
performance. It's overkill too, we don't need these times with
nanosecond precision.

We were using Katip before and it measured event times in milliseconds,
using the auto-update package internally to ensure we'd get the time at
most once every millisecond. We now take the same approach when
measuring the timing of spans.