Drop duplicate query logs
Previously we weren't logging spans themselves, and so we had an
explicit call logging the database query being made. Because we log
database spans now the query information is already being logged, and
these extra logs were printing the information twice. That seems
unnecessary, so we're removing the extra logs.