Don't create a segment for the root span
When sending tracing data to NewRelic we currently log the root span
both as a transaction and as a segment, that is supposed to exactly fill
that root transaction. The idea was that this way we may be able replace
the "C" label in NewRelic with one of our choosing, but it didn't work.

Instead we're going to do the simpler thing now. The root span will be
the NewRelic transaction and it's children will form the first layer of
segments.