Make us of MySQL autocommit
This removes the code that automatically performs a `COMMIT` after every
MySQL query that doesn't take place in a transaction. It shouldn't be
necessary because MySQL does this for us automatically.

Verified by running this query against the prod database, confirming
autocommit is enabled.
  SHOW VARIABLES WHERE Variable_name='autocommit';

We already explicit MySQL transactions in `BEGIN ... COMMIT` blocks, so
there doesn't seem to be any downside to assuming other queries are
transactions.

This change should save us a couple of useless roundtrips to MySQL and a
bit of log spam.