Make `Secret` type more secretive
The `Foldable` instance was providing a function `fold` for unpacking a
secret. We'd like to ensure that `Log.unSecret` is the only way to
unpack a secret, so we have an easy way to audit secrets.

Similarly the `Dhall` instances provided a way to unpack secrets
automatically when sending them into Dhall. This will now need to be
done explicitly.