Simplify `unsafeBulkifyInserts`
The error case we had to handle wasn't used anymore and so could be
removed easily.

That leaves two remaining cases: We're passing in an empty list in which
case the client likely wants to do Nothing, or we pass in at least one
element in which case the client wants to do something with the result.

By letting the client pass in two values for both situations we can
remove the need for an intermediary type.