Rename cherry-core to nri-prelude
We've open-sourced the current version of cherry-core under the name
nri-prelude. This because an open source repository with the name
cherry-core already exists, is a bit different from the version we're
using, and is likely going to remain different for a tad longer to
support experimentation in there.

As a step towards replacing our in-repo `cherry-core` with the
open-sourced version we do the same rename locally we did in the open
source version.