Seperate Redis query building from execution
The goal here is to be able to use the Redis functions both to run
single queries and batched together to read or write multiple things at
the same time.

The change in this commit is internal to the `redis` library, but the
next step will be to push it outward and require consumers of the
`redis` package to use the `query` function explicitly. Once that's done
we'll be set up to add a new `Query` constructor type for transactions.