Use single atomicModifyIORef' in Redis.Mock
This allows us to implement most of the mocking behavior as a pure
function, which is nice.

The real reason for doing this though is that we're about to add
transactions. That means several query operations need to run within a
sinle `atomicModifyIORef'` call. This sets us up for making that change
easier.