Running a Redis transaction is an explicit choice
We were using an imperfect heuristic to determine when to run a
transaction: the existence of an `Apply`, indicating two or more
commands _might_ have been combined.

The implementation of `atomicModify` showed one of possible more
problems with this API. We needed to run a transaction around a single
command to trigger the `WATCH`-ed flags.

Now running a transaction is explicit choice on the part of the user.