Remove concurrent atomicModify test
I don't believe this test can work, maybe the same is true for the
underlying behavior.

I've seen this test fail before when running the specs locally and have
been surprised in the past this didn't seem to affect CI. I think I
understand why now.

All redis tests can be ran against a Real Redis backend or the mock one.
The tests will try to do both by default, but fallback to only running
tests on the mock backend if it can't connect to the real one. I think
this explains why tests were failing locally for me before: they didn't
run against my local Redis cluster, but they did against the mock one.

After modifying the mock Redis implementation to be more like the real
one, now the concurrent tests for the mock backend are failing too.

I believe the problem is that Redis expects `WATCH` to be used on a
connection that's handled in a single-threaded fashion. That is: The
initial `WATCH` of a transaction, the final `EXEC`, and all the commands
in between are executed on a single transaction, with no other logic
using that connection to do its own stuff in between.

The test removed ran a 1000 parallel increment commands on a single key
using a single connection. Redis doesn't guarantee `WATCH/MULTI/EXEC`
atomacity in those circumstances. Suppose all of our threads execute the
following sequence of commands:

    WATCH testkey
    val = GET testkey
    MULTI
    SET testkey (val + 1)
    EXEC

If we interleave commands of two of these threads on a single connection
it can end up like this:

    1. WATCH testkey
    2. WATCH testkey
    1. val = GET testkey
    2. val = GET testkey
    1. MULTI
    1. SET testkey (val + 1)
    2. MULTI
    2. SET testkey (val + 1)
    2. EXEC

The first thread is going to finish between thread two made any changes.
That's good! The second thread should fail because thread 1 modified
`testkey` after thread 2 started watching it. But the `EXEC` sent by
thread 1 removed all watched keys. So by the time `thread 2` multi
starts no keys are being watched, and thread 2 will progress to write
the same incremented value again.

I'm deleting this test because it's failing and I don't see a way to fix
it. I believe `atomicModify` is broken and was before we started making
changes. We should probably remove it and replace it's usage, but that's
out of scope for this commit. Also we need to be really careful using
`WATCH`, because I don't think it's entirely obvious how it works in our
multi-threaded Haskell code.