Report bytes allocated by spans
In case of long running requests this can give us insights in whether
the long running time is explained by the thread being blocked for a
while, or by it doing a ton of computation.

It can also help us get some statistics on what amounts of byte
allocation are reasonable. Once we know this we can automatically kill
threads that go above a certain threshold, preventing a few bad requests
from impacting the overall health of the service.