Fix Redis transactions in cluster mode
We've seen Redis transactions in the quiz-engine service starting to
fail since we enabled the auto-expiry option on the handler. The reason
for this is that the auto-expiry option uses transactions to send the
original requests plus the additional EXPIRE commands in one atomic
package, and we weren't set up to do transactions in a way compatible
with our hedis fork and clustered mode.