Outline alternative test runner
This adds a `TestNew` module that, if we decide to go forward with this
project, will come to replace `Test`. The goal is for the module to get
the exact same API as the current `Test` module, allow us to release it
as a minor change.

The goal of the alternative test module is to drop our dependency on
`tasty` and related libraries. The benefits of this would be:

- Simpler code
- More features (for example: source location for each test!)
- Fewer dependencies!