Catch all Redis exceptions
Currently these are thrown which means our Haskell applications can't
handle them, if they choose to.

This change is prompted by a thrown Redis exeption in the quiz-engine
service. This exception leads to a 500 for the user and bypasses the
logic we have in quiz-engine to ignore any Redis exception. As a result
we had to turn Redis integration off alltogether, preventing us from
gathering data on problems that are happening.