Fix keysTouchedByQuery
`DEL` and `HDEL` do touch keys. The comment we had in place saying this
would result in expire's being added for deleted keys is true, which is
wasteful but does not result in any errors.

What was missing is that the function is also used to figure out which
hashslot to send the `MULTI` command that starts a transaction too. Our
code was not getting any data about keys touched by a transaction
containing only a `DEL` command, because we didn't report on the keys
removed by the `DEL`. This is now fixed.

We've seen some problems related exactly to `DEL`s, that can likely be
explained by this omission.