Undo failing on Redis transactions without a key
In review we realized this change was a bad one, and then I accidentally
pushed it out in a commit later anyway. This undoes that part of the
commit.

The reason we think it's bad idea to fail transactions in this way, is
that there exist legitimate transactions that do not touch any keys,
such as transactions around `PING` commands. The old behavior of just
sending these to any node was the correct behavior, so we reinstate it
here.