Count keys in query once
We use the key list to add one `expire` entry for each key we touch, if
auto-expiry is enabled for the Redis handler. In the old implementation
this could lead to dupplicate expire commands, one for each use of a key
in a transaction. This commit prevents that from happening.