Use Text keys inside of Redis implementation
We'd like to be able to display keys for monitoring purposes, but this
required us to decode them from ByteStrings to Text values. That's
dangerous because parsing might fail even if we don't expect it too.
It's also wasteful because the keys that were originally passed into the
library were already Text values.

By converting our Text keys to ByteStrings in the last possible moment
we're making life easier for ourselves.