Log keys of failing commands
Before this commit we might see an error from Redis telling us that a
`GET` command failed. That's not super useful when we do a lot of them.
This changes it so we log as much information about the commands as is
available and responsible to log, that is everything but the values.