Remove manual boolean to small int conversions
Typically we let our Column and Parameter typeclasses automatically
perform safe conversions from the Postgres/MySql representations of
types to and from the Haskell representations. We made an exceptions for
booleans in MySQL. These are stored as small ints and we were converting
manually from there to the Bool type.

The reason we were doing this was for compatibility with the socket
server implementation of the quiz-engine service. It would decode to
integers, and so our Haskell code had to do the same thing to allow us
to run the same business logic against a socket and a Haskell
implementation.

We killed the socket server a while back though, and since then there's
not really been any reason to continue the conversion dance. This
removes the need for it.