Functions for creating different API's
This makes it so we don't need to expose the Codec machinery to users
outside of the Redis library, cutting back the size of the API.

We loose the ability to define custom codecs outside of the library but
we don't need it right now and have no reason to believe we would in the
future. Should this change we can expose the relevant types and
functions then.