Rename internal Skipped type to NotRan
'skipped' is a bit of an overloaded term. It's used both for tests that
are explicitly skipped using the `skip` function on the test API, but
also for tests that didn't get ran other reasons, such as not being
one of the tests marked with `only`.

Hopefully this rename makes the distinction clearer.