Create MySQL test helper
This commit uses some not-so-nice Haskell to create a nicer experience
of writing Haskell tests using MySQL.

We expose a `MySQL.test` helper that does setup in the right way, and
automatically wraps the body in a test transaction. No need to use the
handler pattern to pass a MySQL connection into individual test cases
(the handler pattern in tests always felt a bit weird to me since tests
are supposed to be isolated).

To make it all possible without paying in performance we need to make
use of `unsafePerformIO`. This particular usage (creating an `MVar`) is
fairly common, but still, would be nicer not to have it.

Overall I feel the upsides outweigh the downsides here.