Always run redis tests with real handler
The reason we weren't doing this before was that we didn't have a real
redis server running in CI. Now that CI is nixified this is easy to
change.

A nice side benefit of this is that we're never running fewer tests then
we might think because we forgot to start the redis server.