Add summary field on tracing span
This is inteded for use in the log explorer tool. Instead of just seeing
that something is 'a test run' or 'a redis query', we'd like to show a
little bit of additional detail. It's hard to summarize the details blob
though, because it's user defined and so we don't know which fields are
the most relevant.