Add support for copying to clipboard
Span details can be huge and can only be shown as JSON documents. Being
able to navigate a JSON document nicely falls (for the moment) outside
the scope of this tool.

This adds a way to easily export spans to the clipboard, so they can be
inspected in more details in an editor, or manipulated using tools like
jq.