Try exponential increase of sample rate
Ironically we're doing this to increase the sample rate, or log fewer
requests. That's not necessarily what we expect with the word
'exponential', but does apply here. The initial sample rate is very low,
like 0.001. If we take linear steps from there to 1 the initial steps
are going to be pretty large. This change should take it slower.