Solve remaining freezes
Not sure why this helped exactly, but this seems to address freezes
while loading spans and pressing buttons simultaneously.

Seems related to both types of events using the same channel to push
messages into the app. This commit makes it so keypresses get dropped if
the event queue is full because of other events (likely data loading).