Log `err` again when using `Expect.Task.succeeds`
Adding the `Show err` results in an `ambigous type err` error when we
pass in a Task with a non-concrete error type, like `Task.succeeds () ::
Task e ()`. We can get rid of these ambiguous errors if we enable the
`ExtendedDefaultRules` extension in our test modules, which maybe we
should make a recommendation for all our test code.