Tweak placement of withFrozenCallStack
This is maybe kinda arbitrary, but I like this style slightly better
because I'm able to explain why it works (tests proof the other style
works though, but I don't have a clue why, maybe because GHC optimizes
the code in a particular way?).

Here's my understanding: The trick with `withFrozenCallStack` is that
you have to pass it the function that isn't supposed to record a
callstack. So:

```hs
2
  |> add 4
  |> withFrozenCallStack
```

Doesn't work, because we're piping the value `6` throuh it, not the
function `add`.

```hs
withFrozenCallStack (add 4 2)
```

I also wouldn't expect to work, because like in the previous example
we're not passing it the function `add`, but rather the result `6` of
calling that function. By the time we evaluate `withFrozenCallStack`
we're too late, `add` has already run.

```hs
withFrozenCallStack add 4 2
```

Looks confusing, but I think kinda makes sense. Read it like so:

```hs
(withFrozenCallStack add) 4 2
```

Or in words `withFrozenCallStack` modifies the function `add`, returning
a new function which takes the same arguments and works the same way,
_except_ it doesn't add a frame to the callstack.

Again, what we had before worked too, but I have no idea why.
Considering the understanding sketched above does reliably get the right
results from `withFrozenCallStack` I think it's worth sticking too!