Don't record traces for fuzz tests, for now
Fuzz tests run a great many times, and each time would record a new
trace. This creates a massive amount of data, most of it unuseful.
Ideally we'd only register the trace of the recording that produced the
error. Until we're able to figure out how to do that it's better to stop
record traces in fuzz tests completely.