Don't pass value through `Expect.andCheck`
The `Expect.andCheck` after removal of `Test.task` and before this
commit was a bit hard to use. Put it in the middle of a test do block
and you'd be fine, but make it the last expectation and you'd run into
trouble. `andCheck` returned an `Expectation' a`, but the expected final
return value is an `Expectation' ()`.

This makes it so `andCheck` returns an `Expectation' ()` so it can be
used to make the last expectation in a test.

We loose the ability to do this:

```hs
test "foo" <| \_ -> do
  res <-
    Task.suceed 4
      |> Expect.andCheck (Expect.equal 4)
  -- do something with `res`
```

But this isn't much of a loss since the removal of `Test.task`, because
now we can also do this:

```hs
test "foo" <| \_ -> do
  res <-
    Task.suceed 4
  Expect.equal 4 res
  -- do something with `res`
```

Which is arguably an improvement.